Application/USB_Serial_Livestream-analysis-Python/requirements.txt
# USB Serial ECG Dashboard - Custom Hardware
# pip install -r requirements.txt

# GUI and visualization
PyQt5>=5.15.0
pyqtgraph>=0.13.0

# USB Serial connectivity
pyserial>=3.5

# Networking (MQTT telemetry)
paho-mqtt>=2.0.0

# Signal processing
scipy>=1.13.0
numpy>=2.0.0
pandas>=2.0.0
neurokit2>=0.2.0
pyhrv>=0.4.1
biosppy>=1.0.0
vital_sqi>=0.1.0
matplotlib>=3.5.0

# HAR Inference (PyTorch models)
# torch is required but version depends on your CUDA setup
# Install separately: pip install torch --index-url https://download.pytorch.org/whl/cu121

# VO2max Neural Network (CardioFitness pretrained model)
tensorflow>=2.12.0
scikit-learn>=1.3.0
joblib>=1.3.0
